+++ src/cmd/build.rs
@@ -247,13 +247,6 @@ fn exec_cargo_for_wasm_target(
 ) -> Result<()> {
     util::assert_channel()?;
 
-    // set linker args via RUSTFLAGS.
-    // Currently will override user defined RUSTFLAGS from .cargo/config. See https://github.com/paritytech/cargo-contract/issues/98.
-    std::env::set_var(
-        "RUSTFLAGS",
-        "-C link-arg=-zstack-size=65536 -C link-arg=--import-memory -Clinker-plugin-lto",
-    );
-
     let cargo_build = |manifest_path: &ManifestPath| {
         let target_dir = &crate_metadata.target_directory;
         let target_dir = format!("--target-dir={}", target_dir.to_string_lossy());
@@ -272,7 +265,11 @@ fn exec_cargo_for_wasm_target(
         } else {
             args.push("-Zbuild-std-features=panic_immediate_abort");
         }
-        util::invoke_cargo(command, &args, manifest_path.directory(), verbosity)?;
+        let env = vec![(
+            "RUSTFLAGS",
+            Some("-C link-arg=-zstack-size=65536 -C link-arg=--import-memory -Clinker-plugin-lto"),
+        )];
+        util::invoke_cargo(command, &args, manifest_path.directory(), verbosity, env)?;
 
         Ok(())
     };
@@ -298,8 +295,112 @@ fn exec_cargo_for_wasm_target(
             .using_temp(cargo_build)?;
     }
 
-    // clear RUSTFLAGS
-    std::env::remove_var("RUSTFLAGS");
+    Ok(())
+}
+
+/// Executes `cargo dylint` with the ink! linting driver that is built during
+/// the `build.rs`.
+///
+/// We create a temporary folder, extract the linting driver there and run
+/// `cargo dylint` with it.
+fn exec_cargo_dylint(crate_metadata: &CrateMetadata, verbosity: Verbosity) -> Result<()> {
+    check_dylint_requirements(crate_metadata.manifest_path.directory())?;
+
+    let tmp_dir = tempfile::Builder::new()
+        .prefix("cargo-contract-dylint_")
+        .tempdir()?;
+    log::debug!("Using temp workspace at '{}'", tmp_dir.path().display());
+
+    let driver = include_bytes!(concat!(env!("OUT_DIR"), "/ink-dylint-driver.zip"));
+    crate::util::unzip(driver, tmp_dir.path().to_path_buf(), None)?;
+
+    let manifest_path = crate_metadata.manifest_path.cargo_arg()?;
+    let args = vec!["--lib", "ink_linting", &manifest_path];
+    let tmp_dir_path = tmp_dir.path().as_os_str().to_string_lossy();
+    let env = vec![
+        ("DYLINT_LIBRARY_PATH", Some(tmp_dir_path.as_ref())),
+        // For tests we need to set the `DYLINT_DRIVER_PATH` to a tmp folder,
+        // otherwise tests running in parallel will try to write to the same
+        // file at the same time which will result in a `Text file busy` error.
+        #[cfg(test)]
+        ("DYLINT_DRIVER_PATH", Some(tmp_dir_path.as_ref())),
+        // We need to remove the `CARGO_TARGET_DIR` environment variable in
+        // case `cargo dylint` is invoked.
+        //
+        // This is because the ink! dylint driver crate found in `dylint` uses a
+        // fixed Rust toolchain via the `ink_linting/rust-toolchain` file. By
+        // removing this env variable we avoid issues with different Rust toolchains
+        // interfering with each other.
+        ("CARGO_TARGET_DIR", None),
+    ];
+    let working_dir = crate_metadata
+        .manifest_path
+        .directory()
+        .unwrap_or_else(|| Path::new("."))
+        .canonicalize()?;
+
+    let verbosity = if verbosity == Verbosity::Verbose {
+        // `dylint` is verbose by default, it doesn't have a `--verbose` argument,
+        Verbosity::Default
+    } else {
+        verbosity
+    };
+    util::invoke_cargo("dylint", &args, Some(working_dir), verbosity, env)?;
+
+    Ok(())
+}
+
+/// Checks if all requirements for `dylint` are installed.
+///
+/// We require only an installed version of `cargo-dylint` here and don't
+/// check for an installed version of `dylint-link`. This is because
+/// `dylint-link` is only required for the `dylint` driver build process
+/// in `build.rs`.
+///
+/// This function takes a `_working_dir` which is only used for unit tests.
+fn check_dylint_requirements(_working_dir: Option<&Path>) -> Result<()> {
+    let execute_cmd = |cmd: &mut Command| {
+        // when testing this function we set the `PATH` to the `working_dir`
+        // so that we can have mocked binaries in there which are executed
+        // instead of the real ones.
+        #[cfg(test)]
+        {
+            let default_dir = PathBuf::from(".");
+            let working_dir = _working_dir.unwrap_or(default_dir.as_path());
+            let path_env = std::env::var("PATH").unwrap();
+            let path_env = format!("{}:{}", working_dir.to_string_lossy(), path_env);
+            cmd.env("PATH", path_env);
+        }
+
+        cmd.stdout(std::process::Stdio::null())
+            .stderr(std::process::Stdio::null())
+            .spawn()
+            .map_err(|err| {
+                log::debug!("Error spawning `{:?}`", cmd);
+                err
+            })?
+            .wait()
+            .map(|res| res.success())
+            .map_err(|err| {
+                log::debug!("Error waiting for `{:?}`: {:?}", cmd, err);
+                err
+            })
+    };
+
+    // when testing this function we should never fall back to a `cargo` specified
+    // in the env variable, as this would mess with the mocked binaries.
+    #[cfg(not(test))]
+    let cargo = std::env::var("CARGO").unwrap_or_else(|_| "cargo".to_string());
+    #[cfg(test)]
+    let cargo = "cargo";
+
+    if !execute_cmd(Command::new(cargo).arg("dylint").arg("--version"))? {
+        anyhow::bail!("cargo-dylint was not found!\n\
+            Make sure it is installed and the binary is in your PATH environment.\n\n\
+            You can install it by executing `cargo install cargo-dylint`."
+            .to_string()
+            .bright_yellow());
+    }
 
     Ok(())
 }
@@ -450,7 +551,7 @@ fn do_optimization(
     let which = which::which("wasm-opt");
     if which.is_err() {
         anyhow::bail!(
-            "wasm-opt not found! Make sure the binary is in your PATH environment.\n\
+            "wasm-opt not found! Make sure the binary is in your PATH environment.\n\n\
             We use this tool to optimize the size of your contract's Wasm binary.\n\n\
             wasm-opt is part of the binaryen package. You can find detailed\n\
             installation instructions on https://github.com/WebAssembly/binaryen#tools.\n\n\
@@ -465,7 +566,7 @@ fn do_optimization(
     }
     let wasm_opt_path = which
         .as_ref()
-        .expect("we just checked if which returned an err; qed")
+        .expect("we just checked if `which` returned an err; qed")
         .as_path();
     log::info!("Path to wasm-opt executable: {}", wasm_opt_path.display());
 
@@ -615,16 +716,15 @@ fn assert_compatible_ink_dependencies(
 ) -> Result<()> {
     for dependency in ["parity-scale-codec", "scale-info"].iter() {
         let args = ["-i", dependency, "--duplicates"];
-        let _ = util::invoke_cargo("tree", &args, manifest_path.directory(), verbosity).map_err(
-            |_| {
+        let _ = util::invoke_cargo("tree", &args, manifest_path.directory(), verbosity, vec![])
+            .map_err(|_| {
                 anyhow::anyhow!(
                     "Mismatching versions of `{}` were found!\n\
                      Please ensure that your contract and your ink! dependencies use a compatible \
                      version of this package.",
                     dependency
                 )
-            },
-        )?;
+            })?;
     }
     Ok(())
 }
@@ -671,6 +771,14 @@ pub(crate) fn execute(args: ExecuteArgs) -> Result<BuildResult> {
             verbosity,
             " {} {}",
             format!("[1/{}]", build_artifact.steps()).bold(),
+            "Checking ink! linting rules".bright_green().bold()
+        );
+        exec_cargo_dylint(&crate_metadata, verbosity)?;
+
+        maybe_println!(
+            verbosity,
+            " {} {}",
+            format!("[2/{}]", build_artifact.steps()).bold(),
             "Building cargo project".bright_green().bold()
         );
         exec_cargo_for_wasm_target(
@@ -685,7 +793,7 @@ pub(crate) fn execute(args: ExecuteArgs) -> Result<BuildResult> {
         maybe_println!(
             verbosity,
             " {} {}",
-            format!("[2/{}]", build_artifact.steps()).bold(),
+            format!("[3/{}]", build_artifact.steps()).bold(),
             "Post processing wasm file".bright_green().bold()
         );
         post_process_wasm(&crate_metadata)?;
@@ -693,7 +801,7 @@ pub(crate) fn execute(args: ExecuteArgs) -> Result<BuildResult> {
         maybe_println!(
             verbosity,
             " {} {}",
-            format!("[3/{}]", build_artifact.steps()).bold(),
+            format!("[4/{}]", build_artifact.steps()).bold(),
             "Optimizing wasm file".bright_green().bold()
         );
         let optimization_result =
@@ -704,6 +812,20 @@ pub(crate) fn execute(args: ExecuteArgs) -> Result<BuildResult> {
 
     let (opt_result, metadata_result) = match build_artifact {
         BuildArtifacts::CheckOnly => {
+            maybe_println!(
+                verbosity,
+                " {} {}",
+                format!("[1/{}]", build_artifact.steps()).bold(),
+                "Checking ink! linting rules".bright_green().bold()
+            );
+            exec_cargo_dylint(&crate_metadata, verbosity)?;
+
+            maybe_println!(
+                verbosity,
+                " {} {}",
+                format!("[2/{}]", build_artifact.steps()).bold(),
+                "Executing `cargo check`".bright_green().bold()
+            );
             exec_cargo_for_wasm_target(
                 &crate_metadata,
                 "check",
@@ -790,6 +912,20 @@ mod tests_ci_only {
             .any(|e| e.name() == "name")
     }
 
+    /// Creates an executable file at `path` with the content `content`.
+    ///
+    /// Currently works only on `unix`.
+    #[cfg(unix)]
+    fn create_executable(path: &Path, content: &str) {
+        {
+            let mut file = std::fs::File::create(&path).unwrap();
+            file.write_all(content.as_bytes())
+                .expect("writing of executable failed");
+        }
+        std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o777))
+            .expect("setting permissions failed");
+    }
+
     /// Creates an executable `wasm-opt-mocked` file which outputs
     /// "wasm-opt version `version`".
     ///
@@ -799,14 +935,8 @@ mod tests_ci_only {
     #[cfg(unix)]
     fn mock_wasm_opt_version(tmp_dir: &Path, version: &str) -> PathBuf {
         let path = tmp_dir.join("wasm-opt-mocked");
-        {
-            let mut file = std::fs::File::create(&path).unwrap();
-            let version = format!("#!/bin/sh\necho \"wasm-opt version {}\"", version);
-            file.write_all(version.as_bytes())
-                .expect("writing wasm-opt-mocked failed");
-        }
-        std::fs::set_permissions(&path, std::fs::Permissions::from_mode(0o777))
-            .expect("setting permissions failed");
+        let content = format!("#!/bin/sh\necho \"wasm-opt version {}\"", version);
+        create_executable(&path, &content);
         path
     }
 
@@ -1270,6 +1400,98 @@ mod tests_ci_only {
 
             // then
             assert!(res.serialize_json().is_ok());
+            Ok(())
+        })
+    }
+
+    // This test has to be ignored until the next ink! rc.
+    // Before that we don't have the `__ink_dylint_…` markers available
+    // to actually run `dylint`.
+    #[test]
+    #[ignore]
+    fn dylint_must_find_issue() {
+        with_new_contract_project(|manifest_path| {
+            // given
+            let contract = r#"
+            #![cfg_attr(not(feature = "std"), no_std)]
+
+            use ink_lang as ink;
+
+            #[ink::contract]
+            mod fail_mapping_01 {
+                use ink_storage::{traits::SpreadAllocate, Mapping};
+
+                #[ink(storage)]
+                #[derive(SpreadAllocate)]
+                pub struct MyContract {
+                    balances: Mapping<AccountId, Balance>,
+                }
+
+                impl MyContract {
+                    #[ink(constructor)]
+                    pub fn new() -> Self {
+                        Self {
+                            balances: Default::default(),
+                        }
+                    }
+
+                    /// Returns the total token supply.
+                    #[ink(message)]
+                    pub fn get(&self) {
+                        // ...
+                    }
+                }
+            }"#;
+            let project_dir = manifest_path.directory().expect("directory must exist");
+            let lib = project_dir.join("lib.rs");
+            std::fs::write(&lib, contract)?;
+
+            let args = crate::cmd::build::ExecuteArgs {
+                manifest_path,
+                build_artifact: BuildArtifacts::CheckOnly,
+                ..Default::default()
+            };
+
+            // when
+            let res = super::execute(args);
+
+            // then
+            match res {
+                Err(err) => {
+                    eprintln!("err: {:?}", err);
+                    assert!(err.to_string().contains(
+                        "help: add an `initialize_contract` function in this constructor"
+                    ));
+                }
+                _ => panic!("build succeeded, but must fail!"),
+            };
+            Ok(())
+        })
+    }
+
+    #[cfg(unix)]
+    #[test]
+    fn missing_cargo_dylint_installation_must_be_detected() {
+        with_new_contract_project(|manifest_path| {
+            // given
+            let manifest_dir = manifest_path.directory().unwrap();
+
+            // mock existing `dylint-link` binary
+            create_executable(&manifest_dir.join("dylint-link"), "#!/bin/sh\nexit 0");
+
+            // mock a non-existing `cargo dylint` installation.
+            create_executable(&manifest_dir.join("cargo"), "#!/bin/sh\nexit 1");
+
+            // when
+            let args = crate::cmd::build::ExecuteArgs {
+                manifest_path,
+                ..Default::default()
+            };
+            let res = super::execute(args).map(|_| ()).unwrap_err();
+
+            // then
+            assert!(format!("{:?}", res).contains("cargo-dylint was not found!"));
+
             Ok(())
         })
     }